refactor(ui): centralize button rendering in shared base
