+++ turbo/packages/ui/src/components/ui/button.tsx
@@ -1,15 +1,13 @@
 import * as React from "react";
-import { Button as ButtonPrimitive } from "@base-ui/react/button";
-import { useRender } from "@base-ui/react/use-render";
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { asChildRender } from "../../lib/base-ui-compat";
 import { cn } from "../../lib/utils";
 import {
+  ButtonBase,
   buttonBaseClassName,
-  ButtonTooltip,
+  type ButtonBaseProps,
   type ButtonTooltipOptions,
-} from "./button-shared";
+} from "./button-base";
 
 const buttonVariants = cva(
   [
@@ -67,99 +65,21 @@ const buttonVariants = cva(
   },
 );
 
-interface ButtonBaseProps
+interface StyledButtonProps
   extends
-    Omit<ButtonPrimitive.Props, "className" | "render">,
-    VariantProps<typeof buttonVariants> {
-  asChild?: boolean;
-  className?: string;
-  render?: ButtonPrimitive.Props["render"];
-}
+    Omit<ButtonBaseProps, "showTooltip" | "tooltipFullWidth">,
+    VariantProps<typeof buttonVariants> {}
 
-export type ButtonProps = ButtonBaseProps & ButtonTooltipOptions;
-
-interface ButtonAsChildProps {
-  children: React.ReactNode;
-  className: string;
-  props: Omit<
-    ButtonPrimitive.Props,
-    "children" | "className" | "nativeButton" | "ref" | "render"
-  >;
-  ref: React.ForwardedRef<HTMLElement>;
-}
-
-function ButtonAsChild({
-  children,
-  className,
-  props,
-  ref,
-}: ButtonAsChildProps) {
-  return useRender({
-    defaultTagName: "button",
-    props: {
-      ...props,
-      className,
-      "data-slot": "button",
-    },
-    ref,
-    render: asChildRender(children),
-  });
-}
+export type ButtonProps = StyledButtonProps & ButtonTooltipOptions;
 
 const Button = React.forwardRef<HTMLElement, ButtonProps>(
-  (
-    {
-      asChild = false,
-      children,
-      className,
-      iconSize,
-      nativeButton,
-      render,
-      showTooltip = false,
-      size,
-      variant,
-      ...props
-    },
-    ref,
-  ) => {
-    const resolvedClassName = cn(
-      buttonVariants({ variant, size, iconSize, className }),
-    );
-    const { title: _title, ...propsWithoutTitle } = props;
-    const buttonProps = showTooltip ? propsWithoutTitle : props;
-
-    const button = asChild ? (
-      <ButtonAsChild
-        className={resolvedClassName}
-        props={buttonProps}
-        ref={ref}
-      >
-        {children}
-      </ButtonAsChild>
-    ) : (
-      <ButtonPrimitive
-        className={resolvedClassName}
-        data-slot="button"
-        nativeButton={nativeButton}
-        ref={ref}
-        render={render}
-        {...buttonProps}
-      >
-        {children}
-      </ButtonPrimitive>
-    );
-
-    if (!showTooltip) {
-      return button;
-    }
-
+  ({ className, iconSize, size, variant, ...props }, ref) => {
     return (
-      <ButtonTooltip
-        disabled={buttonProps.disabled}
-        label={buttonProps["aria-label"]}
-      >
-        {button}
-      </ButtonTooltip>
+      <ButtonBase
+        {...props}
+        className={cn(buttonVariants({ variant, size, iconSize, className }))}
+        ref={ref}
+      />
     );
   },
 );
